+++ CHANGELOG.md
@@ -5,8 +5,16 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+
+## [1.0.0] - 2021-10-21
 ### Added
+- settings state functionality
+- remove moves file
 - gitkeep for dist folder(for github pages)
+### Changed
+- mapping functionality
+### Removed
+- log in actions
 
 ## [0.2.0] - 2021-10-21
 ### Added
@@ -21,6 +29,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - changelog
 - old app version
 
-[Unreleased]: https://github.com/ArtemNikolaev/command-vs-strategy/compare/v0.2.0...HEAD
+[Unreleased]: https://github.com/ArtemNikolaev/command-vs-strategy/compare/v1.0.0...HEAD
+[Unreleased]: https://github.com/ArtemNikolaev/command-vs-strategy/compare/v0.2.0...v1.0.0
 [0.2.0]: https://github.com/ArtemNikolaev/command-vs-strategy/compare/v0.1.0...v0.2.0
 [0.1.0]: https://github.com/ArtemNikolaev/command-vs-strategy/releases/tag/v0.1.0